requirements.txt
streamlit==1.32.2
deep_translator==1.11.4
PyPDF2==3.0.1
google-generativeai==0.4.1